+++ backend/api/Controllers/InspectionAreaController.cs
