+++ backend/api/Controllers/Models/InspectionAreaResponse.cs
