+++ backend/api/Services/InspectionAreaService.cs
