+++ frontend/src/components/Pages/InspectionPage/InspectionAreaCards.tsx
