+++ frontend/src/components/Pages/InspectionPage/InspectionSection.tsx
@@ -1,12 +1,12 @@
 import { useState, useEffect } from 'react'
-import { InspectionArea } from 'models/Deck'
+import { InspectionArea } from 'models/InspectionArea'
 import { useInstallationContext } from 'components/Contexts/InstallationContext'
 import { MissionDefinition } from 'models/MissionDefinition'
 import { ScheduleMissionDialog } from './ScheduleMissionDialogs'
 import { getInspectionDeadline } from 'utils/StringFormatting'
 import { InspectionTable } from './InspectionTable'
 import { StyledDict, compareInspections } from './InspectionUtilities'
-import { InspectionAreaCards } from './DeckCards'
+import { InspectionAreaCards } from './InspectionAreaCards'
 import { Area } from 'models/Area'
 import { useMissionsContext } from 'components/Contexts/MissionRunsContext'
 import { useMissionDefinitionsContext } from 'components/Contexts/MissionDefinitionsContext'
