+++ frontend/src/models/InspectionArea.ts
