+++ frontend/src/utils/InspectionAreaMapView.tsx
